:recycle: Move repeater helper
