+++ corehq/motech/openmrs/repeater_helpers.py
@@ -10,7 +10,6 @@ from requests import RequestException
 from urllib3.exceptions import HTTPError
 
 from casexml.apps.case.mock import CaseBlock
-from casexml.apps.case.xform import extract_case_blocks
 
 from corehq.apps.case_importer import util as importer_util
 from corehq.apps.case_importer.const import LookupErrors
@@ -33,7 +32,6 @@ from corehq.motech.openmrs.exceptions import (
 from corehq.motech.openmrs.finders import PatientFinder
 from corehq.motech.requests import Requests
 from corehq.motech.value_source import (
-    CaseTriggerInfo,
     as_value_source,
     get_ancestor_location_metadata_value,
     get_case_location,
@@ -432,33 +430,6 @@ def delete_case_property(
         case_block_kwargs = {"update": {case_property: None}}
     case_block = CaseBlock(case_id=case_id, create=False, **case_block_kwargs)
     submit_case_blocks([case_block.as_text()], domain, xmlns=XMLNS_OPENMRS)
-
-
-def get_relevant_case_updates_from_form_json(domain, form_json, case_types, extra_fields,
-                                             form_question_values=None):
-    result = []
-    case_blocks = extract_case_blocks(form_json)
-    cases = CaseAccessors(domain).get_cases(
-        [case_block['@case_id'] for case_block in case_blocks], ordered=True)
-    for case, case_block in zip(cases, case_blocks):
-        assert case_block['@case_id'] == case.case_id
-        if not case_types or case.type in case_types:
-            case_create = case_block.get('create') or {}
-            case_update = case_block.get('update') or {}
-            result.append(CaseTriggerInfo(
-                domain=domain,
-                case_id=case_block['@case_id'],
-                type=case.type,
-                name=case.name,
-                owner_id=case.owner_id,
-                modified_by=case.modified_by,
-                updates={**case_create, **case_update},
-                created='create' in case_block,
-                closed='close' in case_block,
-                extra_fields={field: case.get_case_property(field) for field in extra_fields},
-                form_question_values=form_question_values or {},
-            ))
-    return result
 
 
 def get_export_data(config, properties, case_trigger_info):
